account/requirements.txt
# server
fastapi[all]==0.89.1
Jinja2
gunicorn
sentry-sdk
requests

# code formatting
flake8
black

# database
databases[postgresql]
aerich
redis
tortoise-orm[asyncpg]

# unit/integration tests
pytest
pytest-asyncio
httpx
asgi-lifespan
asynctest

# Password Hashing/ Authentication
passlib[bcrypt]
python-jose[cryptography]

# AWS
boto3
